fix integration tests for docker.local
