+++ code/go/0chain.net/blobbercore/handler/handler_integration_tests.go
@@ -14,6 +14,7 @@ import (
 	"os"
 	"runtime/pprof"
 
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/blobberhttp"
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/config"
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/datastore"
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/stats"
@@ -25,6 +26,8 @@ import (
 
 	// integration tests RPC control
 	crpc "github.com/0chain/blobber/code/go/0chain.net/conductor/conductrpc"
+
+	. "github.com/0chain/blobber/code/go/0chain.net/core/logging"
 )
 
 var storageHandler StorageHandler
@@ -40,9 +43,9 @@ func SetupHandlers(r *mux.Router) {
 	r.HandleFunc("/v1/file/download/{allocation}", common.UserRateLimit(common.ToByteStream(WithConnection(DownloadHandler))))
 	r.HandleFunc("/v1/file/rename/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithConnection(RenameHandler))))
 	r.HandleFunc("/v1/file/copy/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithConnection(CopyHandler))))
-	r.HandleFunc("/v1/file/attributes/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithConnection(UpdateObjectAttributes))))
+	r.HandleFunc("/v1/file/attributes/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithConnection(UpdateAttributesHandler))))
 
-	r.HandleFunc("/v1/connection/commit/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithConnection(CommitHandler))))
+	r.HandleFunc("/v1/connection/commit/{allocation}", common.UserRateLimit(common.ToStatusCode(WithStatusConnection(CommitHandler))))
 	r.HandleFunc("/v1/file/commitmetatxn/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithConnection(CommitMetaTxnHandler))))
 
 	//object info related apis
@@ -52,7 +55,7 @@ func SetupHandlers(r *mux.Router) {
 	r.HandleFunc("/v1/file/list/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithReadOnlyConnection(ListHandler))))
 	r.HandleFunc("/v1/file/objectpath/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithReadOnlyConnection(ObjectPathHandler))))
 	r.HandleFunc("/v1/file/referencepath/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithReadOnlyConnection(ReferencePathHandler))))
-	r.HandleFunc("/v1/file/objecttree/{allocation}", common.UserRateLimit(common.ToJSONResponse(WithReadOnlyConnection(ObjectTreeHandler))))
+	r.HandleFunc("/v1/file/objecttree/{allocation}", common.UserRateLimit(common.ToStatusCode(WithStatusReadOnlyConnection(ObjectTreeHandler))))
 
 	//admin related
 	r.HandleFunc("/_debug", common.UserRateLimit(common.ToJSONResponse(DumpGoRoutines)))
@@ -188,7 +191,7 @@ func DownloadHandler(ctx context.Context, r *http.Request) (interface{}, error) 
 
 	var state = crpc.Client().State()
 	if state.StorageTree.IsBad(state, node.Self.ID) {
-		dr := response.(*DownloadResponse)
+		dr := response.(*blobberhttp.DownloadResponse)
 		dr.Path = "/injection/" + dr.Path
 	}
 
@@ -211,20 +214,19 @@ func ListHandler(ctx context.Context, r *http.Request) (interface{}, error) {
 	return response, nil
 }
 
-/*CommitHandler is the handler to respond to upload requests fro clients*/
-func CommitHandler(ctx context.Context, r *http.Request) (interface{}, error) {
-	ctx = setupHandlerContext(ctx, r)
+func CommitHandler(ctx context.Context, r *http.Request) (interface{}, int, error) {
+	resp, status, err := commitHandler(ctx, r)
 
-	response, err := storageHandler.CommitWrite(ctx, r)
 	if err != nil {
-		return nil, err
+		return nil, status, err
 	}
 
 	var state = crpc.Client().State()
 	if state.StorageTree.IsBad(state, node.Self.ID) {
+		response := resp.(*blobberhttp.CommitResult)
 		response.AllocationRoot = revertString(response.AllocationRoot)
 	}
-	return response, nil
+	return resp, http.StatusOK, nil
 }
 
 func ReferencePathHandler(ctx context.Context, r *http.Request) (interface{}, error) {
@@ -264,22 +266,21 @@ func ObjectPathHandler(ctx context.Context, r *http.Request) (interface{}, error
 	return response, nil
 }
 
-func ObjectTreeHandler(ctx context.Context, r *http.Request) (interface{}, error) {
-	ctx = setupHandlerContext(ctx, r)
-
-	response, err := storageHandler.GetObjectTree(ctx, r)
+func ObjectTreeHandler(ctx context.Context, r *http.Request) (interface{}, int, error) {
+	resp, status, err := objectTreeHandler(ctx, r)
 	if err != nil {
-		return nil, err
+		return nil, status, err
 	}
 
 	var state = crpc.Client().State()
 	if state.StorageTree.IsBad(state, node.Self.ID) {
+		response := resp.(*blobberhttp.ReferencePathResult)
 		if len(response.List) > 0 {
 			response.List = append(response.List, response.List[0])
 		}
 	}
 
-	return response, nil
+	return resp, http.StatusOK, nil
 }
 
 func RenameHandler(ctx context.Context, r *http.Request) (interface{}, error) {
@@ -291,7 +292,7 @@ func RenameHandler(ctx context.Context, r *http.Request) (interface{}, error) {
 
 	var state = crpc.Client().State()
 	if state.StorageTree.IsBad(state, node.Self.ID) {
-		ur := response.(*UploadResult)
+		ur := response.(*blobberhttp.UploadResult)
 		ur.Filename = "/injected/" + ur.Filename
 	}
 
@@ -307,7 +308,7 @@ func CopyHandler(ctx context.Context, r *http.Request) (interface{}, error) {
 
 	var state = crpc.Client().State()
 	if state.StorageTree.IsBad(state, node.Self.ID) {
-		ur := response.(*UploadResult)
+		ur := response.(*blobberhttp.UploadResult)
 		ur.Filename = "/injected/" + ur.Filename
 	}
 
@@ -325,6 +326,22 @@ func UploadHandler(ctx context.Context, r *http.Request) (interface{}, error) {
 	var state = crpc.Client().State()
 	if state.StorageTree.IsBad(state, node.Self.ID) {
 		response.Filename = "/injected/" + response.Filename
+	}
+
+	return response, nil
+}
+
+func UpdateAttributesHandler(ctx context.Context, r *http.Request) (interface{}, error) {
+	ctx = setupHandlerContext(ctx, r)
+	response, err := storageHandler.UpdateObjectAttributes(ctx, r)
+	if err != nil {
+		return nil, err
+	}
+
+	var state = crpc.Client().State()
+	if state.StorageTree.IsBad(state, node.Self.ID) {
+		ur := response.(*blobberhttp.UploadResult)
+		ur.Filename = "/injected/" + ur.Filename
 	}
 
 	return response, nil
