+++ docker.local/IntegrationTestsBlobberDockerfile
@@ -7,13 +7,13 @@ ENV GO111MODULE=on
 
 # Download the dependencies:
 # Will be cached if we don't change mod/sum files
-COPY ./code/go/0chain.net/go.mod          ./code/go/0chain.net/go.sum          $SRC_DIR/go/0chain.net/
-RUN cd $SRC_DIR/go/0chain.net && go mod download
+COPY ./go.mod          ./go.sum          $SRC_DIR/
+RUN cd $SRC_DIR && go mod download
 
 #Add the source code
-ADD ./code/go/0chain.net $SRC_DIR/go/0chain.net
+ADD ./code/go/0chain.net $SRC_DIR/code/go/0chain.net
 
-WORKDIR $SRC_DIR/go/0chain.net/blobber
+WORKDIR $SRC_DIR/code/go/0chain.net/blobber
 
 RUN CGO_ENABLED=1 go build -v -tags "bn256 development integration_tests" -ldflags "-X github.com/0chain/blobber/code/go/0chain.net/core/build.BuildTag=$GIT_COMMIT"
 
@@ -25,4 +25,4 @@ COPY --from=blobber_build  /usr/local/lib/libmcl*.so \
                         /usr/local/lib/
 ENV APP_DIR=/blobber
 WORKDIR $APP_DIR
-COPY --from=blobber_build $APP_DIR/go/0chain.net/blobber/blobber $APP_DIR/bin/blobber
+COPY --from=blobber_build $APP_DIR/code/go/0chain.net/blobber/blobber $APP_DIR/bin/blobber
